Add files via upload
Created a web page with my favorite recipes! The recipe book has a table of contents linking to each recipe below, using internal links. Each recipe has a table for the ingredients, a list of details, and a list and/or paragraphs of the steps required. If you use recipes from the internet, be sure to add a link to the website.